+++ NEWS.md
@@ -1,4 +1,4 @@
-# pipapi (development version)
+# pipapi 1.2.1
 
 ## New features
 - New indicators available. Optional with `additional_ind = TRUE`
@@ -10,9 +10,11 @@
 - A new `/ui_aux` endpoint has been created with the opposite behavior: tables are
 returned in wide format by default.
 - `/pip-info` returns additional information
+- New "etag" and "max-age" headers returned by the API to facilitate caching of API responses
 
 ## Enhancements
-
+- pip-grp and cp-chart endpoint small optimization
+- support to arrow/feather serialization with `format=arrow`
 - [Implement checks for long_format parameter in API filter](https://github.com/PIP-Technical-Team/pipapi/pull/332)
 
 ## Bug fixes